reduce likelihood of hitting a lambda timeout (#7)
This change should reduce the likelihood that users (SEs, etc.) have to troubleshoot
Lambda timeouts.

This change does that by adding pagination logic to the Lambda. Instead of creating all
subscriptions in a single invocation, the Lambda sends a custom Eventbridge event which
triggers another Lambda invocation.

So, Lambda invocations are shorter and timeouts are far less likely to be hit.

In addition, this PR adds essential features like the cloudformation clone, changelog, and
a memory parameter.