feat: ignore_delete_errors variable
Currently, if the Python code returns an exception, the user needs to
delete the CloudFormation Stack through the UI or CLI. This change
makes it so that a variable can be changed to ignore deletion issues.

Specifically, this addresses the case where
log_group_matches is invalid regex and the user wants to change it.